move  scala-2.12 module
