+++ linkis-engineconn-plugins/spark/scala-2.12/pom.xml
@@ -21,7 +21,7 @@
     <groupId>org.apache.linkis</groupId>
     <artifactId>linkis</artifactId>
     <version>${revision}</version>
-    <relativePath>../../pom.xml</relativePath>
+    <relativePath>../../../pom.xml</relativePath>
   </parent>
 
   <artifactId>linkis-engineplugin-scala-2.12</artifactId>
