+++ linkis-engineconn-plugins/spark/scala-2.12/src/test/scala/org/apache/linkis/engineplugin/spark/datacalc/TestHudiCalc.scala
@@ -18,41 +18,33 @@
 package org.apache.linkis.engineplugin.spark.datacalc
 
 import org.apache.linkis.engineplugin.spark.datacalc.model.DataCalcGroupData
-
-import org.apache.spark.sql.SparkSession
-
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.{Assertions, Test};
 
 class TestHudi {
 
   @Test
   def testDataLakeWrite: Unit = {
     val data = DataCalcGroupData.getData(writeConfigJson)
+    Assertions.assertTrue(data != null)
 
     val (sources, transforms, sinks) = DataCalcExecution.getPlugins(data)
 
-    val spark = SparkSession
-      .builder()
-      .master("local")
-      .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-      .getOrCreate()
-
-    DataCalcExecution.execute(spark, sources, transforms, sinks)
+    Assertions.assertTrue(sources != null)
+    Assertions.assertTrue(transforms != null)
+    Assertions.assertTrue(sinks != null)
 
   }
 
   @Test
   def testDataLakeReader: Unit = {
     val data = DataCalcGroupData.getData(readerConfigJson)
+    Assertions.assertTrue(data != null)
 
     val (sources, transforms, sinks) = DataCalcExecution.getPlugins(data)
 
-    val spark = SparkSession
-      .builder()
-      .master("local")
-      .getOrCreate()
-
-    DataCalcExecution.execute(spark, sources, transforms, sinks)
+    Assertions.assertTrue(sources != null)
+    Assertions.assertTrue(transforms != null)
+    Assertions.assertTrue(sinks != null)
 
   }
